# Conflicts:
#	.gitignore
#	include/cn-cbor/cn-cbor.h
#	src/cn-cbor.c
#	src/cn-encoder.c

+++ src/cn-encoder.c
@@ -21,6 +21,7 @@ extern "C" {
 #include <stdbool.h>
 #include <assert.h>
 
+#include "dll-export.h"
 #include "cn-cbor/cn-cbor.h"
 #include "cbor.h"
 
@@ -97,7 +98,7 @@ static void _write_positive(cn_write_state *ws, cn_cbor_type typ, uint64_t val) 
 
   if (val < 24) {
     ensure_writable(1);
-    write_byte(ib | val);
+    write_byte(ib | (uint8_t) val);
   } else if (val < 256) {
     ensure_writable(2);
     write_byte(ib | 24);
